refactor: DDD 패키지 구조 적용 및 불필요한 패턴 제거
