+++ src/main/java/com/one/domain/category/application/UserBigCategorySaveService.java
@@ -1,0 +1,17 @@
+package com.one.domain.category.application;
+
+import com.one.domain.category.domain.UserBigCategory;
+import com.one.domain.category.domain.CategoryMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserBigCategorySaveService {
+
+    private final CategoryMapper categoryMapper;
+
+    public void save(final UserBigCategory userBigCategory) {
+        categoryMapper.saveUserBigCategory(userBigCategory);
+    }
+}
