--- src/main/java/com/one/domain/category/application/UserBigCategoryService.java
@@ -1,7 +1,0 @@
-package com.one.domain.category.application;
-
-import com.one.domain.category.dto.UserBigCategorySaveRequestDto;
-
-public interface UserBigCategoryService {
-    void save(UserBigCategorySaveRequestDto userBigCategorySaveRequestDto);
-}
