+++ src/main/java/com/one/domain/category/domain/CategoryMapper.java
@@ -1,0 +1,10 @@
+package com.one.domain.category.domain;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface CategoryMapper {
+    void saveUserBigCategory(UserBigCategory userBigCategory);
+
+    void deleteAll();
+}
