+++ src/main/java/com/one/domain/category/domain/UserBigCategory.java
@@ -1,0 +1,4 @@
+package com.one.domain.category.domain;
+
+public record UserBigCategory(int userId, int bigCategoryId) {
+}
