--- src/main/java/com/one/domain/category/dto/UserBigCategorySaveRequestDto.java
@@ -1,4 +1,0 @@
-package com.one.domain.category.dto;
-
-public record UserBigCategorySaveRequestDto(int userId, int bigCategoryId) {
-}
