+++ src/main/java/com/one/domain/file/application/FileManagementService.java
@@ -1,11 +1,26 @@
 package com.one.domain.file.application;
 
 import com.one.domain.file.code.ImageFileType;
+import com.one.domain.file.domain.FileManager;
+import com.one.domain.file.dto.ImageFileSaveRequestDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.env.Environment;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
+@Service
+@RequiredArgsConstructor
+public class FileManagementService {
 
-public interface FileManagementService {
-    int upload(MultipartFile multipartFile, ImageFileType imageFileType) throws IOException;
-    void download(int imageFileId);
+    private final FileManager fileManager;
+    private final Environment environment;
+
+    @Transactional
+    public int upload(final MultipartFile multipartFile, final ImageFileType imageFileType) {
+        final ImageFileSaveRequestDto imageFileSaveRequestDto = ImageFileSaveRequestDto.of(environment.getProperty("file.dir"), multipartFile, imageFileType);
+        final int id = fileManager.upload(imageFileSaveRequestDto);
+        fileManager.save(multipartFile, imageFileSaveRequestDto.getPath());
+        return id;
+    }
 }
