+++ src/main/java/com/one/domain/file/domain/ImageFile.java
@@ -1,0 +1,6 @@
+package com.one.domain.file.domain;
+
+import java.time.LocalDateTime;
+
+public record ImageFile(Integer id, String path, String name, int contentType, LocalDateTime createTime, LocalDateTime updateTime) {
+}
