+++ src/main/java/com/one/domain/file/domain/ImageFileMapper.java
@@ -1,0 +1,16 @@
+package com.one.domain.file.domain;
+
+import com.one.domain.file.dto.ImageFileSaveRequestDto;
+import com.one.domain.file.dto.ImageFileSaveRequestRecord;
+import org.apache.ibatis.annotations.Mapper;
+
+import java.util.Optional;
+
+@Mapper
+public interface ImageFileMapper {
+    int saveImageFile(ImageFileSaveRequestDto imageFileSaveRequestDto);
+
+    int saveImageFileByRecord(ImageFileSaveRequestRecord imageFileSaveRequestRecord);
+
+    Optional<ImageFile> findImageFileById(int id);
+}
