+++ src/main/java/com/one/domain/file/dto/ImageFileSaveRequestRecord.java
@@ -1,0 +1,4 @@
+package com.one.domain.file.dto;
+
+public record ImageFileSaveRequestRecord(Integer id, String path, String name, int contentType) {
+}
