+++ src/main/java/com/one/domain/sms/domain/SmsAuthenticationRepository.java
@@ -1,0 +1,7 @@
+package com.one.domain.sms.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface SmsAuthenticationRepository extends CrudRepository<SmsAuthentication, String> {
+
+}
