--- src/main/java/com/one/domain/sms/model/SmsAuthentication.java
@@ -1,9 +1,0 @@
-package com.one.domain.sms.model;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.redis.core.RedisHash;
-
-@RedisHash(value = "sms", timeToLive = 60 * 3)
-public record SmsAuthentication(@Id String phoneNumber, String authenticationNumber) {
-
-}
