--- src/main/java/com/one/domain/sms/repository/SmsAuthenticationRepository.java
@@ -1,8 +1,0 @@
-package com.one.domain.sms.repository;
-
-import com.one.domain.sms.model.SmsAuthentication;
-import org.springframework.data.repository.CrudRepository;
-
-public interface SmsAuthenticationRepository extends CrudRepository<SmsAuthentication, String> {
-
-}
