--- src/main/java/com/one/domain/user/application/UserSaveService.java
@@ -1,7 +1,0 @@
-package com.one.domain.user.application;
-
-import com.one.domain.user.dto.UserSaveRequestDto;
-
-public interface UserSaveService {
-    int save(UserSaveRequestDto userSaveRequestDto);
-}
