--- src/main/resources/com/one/domain/file/mapper/ImageFileMapper.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.one.domain.file.mapper.ImageFileMapper">
-    <insert id="saveImageFile" parameterType="com.one.domain.file.dto.ImageFileSaveRequestDto" useGeneratedKeys="true" keyProperty="id">
-        insert
-          into image_file
-               (path, name, content_type, create_time, update_time)
-        values (#{path}, #{name}, #{contentType}, now(), now())
-    </insert>
-
-    <select id="findImageFileById" resultType="com.one.domain.file.model.ImageFile">
-        select id, path, name, content_type, create_time, update_time
-        from image_file
-        where id = #{id}
-    </select>
-</mapper>
