+++ src/test/java/com/one/domain/file/mapper/ImageFileMapperTest.java
@@ -1,7 +1,9 @@
 package com.one.domain.file.mapper;
 
+import com.one.domain.file.domain.ImageFileMapper;
 import com.one.domain.file.dto.ImageFileSaveRequestDto;
-import com.one.domain.file.model.ImageFile;
+import com.one.domain.file.dto.ImageFileSaveRequestRecord;
+import com.one.domain.file.domain.ImageFile;
 import com.one.domain.file.code.ImageFileType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -20,7 +22,7 @@ class ImageFileMapperTest {
     @Test
     @DisplayName("이미지 파일 정보를 DB에 저장하고 조회한다")
     public void test() {
-        final ImageFileSaveRequestDto imageFileSaveRequestDto = new ImageFileSaveRequestDto(null, "testPath", "testName", ImageFileType.A);
+        final ImageFileSaveRequestDto imageFileSaveRequestDto = new ImageFileSaveRequestDto(null, "testPath", "testName", ImageFileType.USER.getValue());
         imageFileMapper.saveImageFile(imageFileSaveRequestDto);
         Optional<ImageFile> imageFile = imageFileMapper.findImageFileById(imageFileSaveRequestDto.getId());
         Assertions.assertThat(imageFile.get().path()).isEqualTo(imageFileSaveRequestDto.getPath());
@@ -28,4 +30,9 @@ class ImageFileMapperTest {
         Assertions.assertThat(imageFile.get().contentType()).isEqualTo(imageFileSaveRequestDto.getContentType());
     }
 
+    @Test
+    public void test2() {
+        final ImageFileSaveRequestRecord imageFileSaveRequestRecord = new ImageFileSaveRequestRecord(null, "testPath", "testName", ImageFileType.USER.getValue());
+        imageFileMapper.saveImageFileByRecord(imageFileSaveRequestRecord);
+    }
 }
